Extract base props from cql-sdk which removes some of the unnecessary properties for projects like tests
