+++ Cql/CoreTests/Fhir/FhirCqlContextCreateTests.cs
@@ -1,18 +1,12 @@
 ﻿#nullable enable
-using System;
-using System.Collections.Generic;
-using System.Dynamic;
-using System.Linq;
-using System.Reflection;
 using FluentAssertions;
 using Hl7.Cql.Fhir;
 using Hl7.Cql.Primitives;
 using Hl7.Cql.Runtime;
-using Hl7.Fhir.Model;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using TypeConverter = Hl7.Cql.Conversion.TypeConverter;
 
 namespace CoreTests.Fhir;
+
 
 [TestClass]
 public class FhirCqlContextCreateTests
@@ -79,82 +73,16 @@ public class FhirCqlContextCreateTests
         typeConverterLRUCache1.Should().BeSameAs(typeConverterLRUCache2);
     }
 
-    #region Helpers
-
-    private static CqlContext CreateCqlContext(
+    protected static CqlContext CreateCqlContext(
         FhirCqlContextOptions? options = null)
     {
         var context = FhirCqlContext.CreateContext(options: options);
         context.Should().NotBeNull();
         return context;
     }
-
-    #endregion
 }
 
 file static class Extensions
 {
     public static CqlOperators GetCqlOperators(this CqlContext context) => context.Operators.Should().BeOfType<CqlOperators>().Subject;
-
-    public static dynamic AccessMembersDynamically(this object? obj) => obj is null ? DynamicMemberAccessor.Null : new DynamicMemberAccessor(obj);
-
-    public static Maybe<LRUCache<CqlDateTime>> GetLRUCache(this TypeConverter typeConverter)
-    {
-        Dictionary<Type, Dictionary<Type, Func<object, object>>> converters = typeConverter.AccessMembersDynamically()._converters;
-        Dictionary<Type, Func<object, object>> fhirDateTimeConverters = converters[typeof(FhirDateTime)];
-        Func<object, object> firstConversion = fhirDateTimeConverters.Values.First();
-        LRUCache<CqlDateTime>? lruCache = firstConversion.Target!.AccessMembersDynamically().conversion.Target.dateTimes;
-        return lruCache;
-    }
-
-    public static int GetCapacity(this LRUCache<CqlDateTime> lruCache)
-    {
-        int capacity = lruCache.AccessMembersDynamically().Capacity;
-        return capacity;
-    }
-}
-
-file readonly record struct Maybe<T>
-{
-    private readonly T _value;
-
-    private Maybe(T Value)
-    {
-        this._value = Value;
-    }
-
-    public T Value => HasValue ? _value : throw new InvalidOperationException("No value");
-
-    public bool HasValue { get; } = true;
-
-    public static implicit operator Maybe<T>(T? value) => value is null ? default : new(value);
-}
-
-file class DynamicMemberAccessor(object? value) : DynamicObject
-{
-    public static readonly DynamicMemberAccessor Null = new(null);
-    private object? Value { get; } = value;
-    private Type? Type { get; } = value?.GetType();
-
-    public override bool TryGetMember(GetMemberBinder binder, out object? result)
-    {
-        switch (Type?.GetMember(binder.Name, BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic))
-        {
-            case [FieldInfo field]:
-                result =field.GetValue(Value).AccessMembersDynamically();
-                return true;
-            case [PropertyInfo prop] when prop.GetIndexParameters() is []:
-                result = prop.GetValue(Value).AccessMembersDynamically();
-                return true;
-            default:
-                result = Null;
-                return true;
-        }
-    }
-
-    public override bool TryConvert(ConvertBinder binder, out object? result)
-    {
-        result = Value;
-        return true;
-    }
 }
