+++ cql-base.props
@@ -1,0 +1,89 @@
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+	</PropertyGroup>
+
+	<PropertyGroup>
+		<!-- Keep this in sync between cql-sdk.props and cql-demo.props -->
+		<FirelySdkVersion>5.10.3</FirelySdkVersion>
+	</PropertyGroup>
+
+	<!-- Solution-wide properties for NuGet packaging -->
+	<PropertyGroup>
+		<!--
+			Instructions when updating VersionPrefix/VersionSuffix : https://github.com/FirelyTeam/firely-cql-sdk/wiki/Creating-Tags-and-Releases
+			-->
+		<!-- <VersionPrefix>2.0.4</VersionPrefix> -->
+		<!-- <VersionSuffix>alpha</VersionSuffix> -->
+		<!-- <Authors>NCQA, Firely (info@fire.ly) and contributors</Authors> -->
+		<!-- <Copyright>Copyright 2013-2024 NCQA, Firely. Contains materials (C) HL7 International</Copyright> -->
+		<!-- <PackageProjectUrl>https://github.com/FirelyTeam/firely-cql-sdk</PackageProjectUrl> -->
+		<!-- <RepositoryUrl>https://github.com/FirelyTeam/firely-cql-sdk</RepositoryUrl> -->
+		<!-- <RepositoryType>git</RepositoryType> -->
+		<!-- <PackageReleaseNotes>See https://github.com/FirelyTeam/firely-cql-sdk/releases</PackageReleaseNotes> -->
+		<!-- <PackageLicenseExpression>BSD-3-Clause</PackageLicenseExpression> -->
+		<!-- <PackageReadmeFile>README.md</PackageReadmeFile> -->
+		<!-- <PublishRepositoryUrl>true</PublishRepositoryUrl> -->
+		<!-- <EmbedUntrackedSources>true</EmbedUntrackedSources> -->
+		<!-- Use CqlSolutionDir instead of SolutionDir, because this file is used in different solutions across repos -->
+		<CqlSolutionDir>$(MSBuildThisFileDirectory)</CqlSolutionDir>
+	</PropertyGroup>
+
+	<!-- <ItemGroup> -->
+		<!-- <None Include="..\..\README.md" Pack="true" PackagePath=""/> -->
+	<!-- </ItemGroup> -->
+
+	<!-- <ItemGroup> -->
+		<!-- <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All"/> -->
+	<!-- </ItemGroup> -->
+
+	<!-- Compiler settings -->
+	<PropertyGroup>
+		<Nullable>enable</Nullable>
+		<LangVersion>12</LangVersion>
+		<!-- <GenerateDocumentationFile>True</GenerateDocumentationFile> -->
+		<!-- <Configurations>Debug;Release;FullDebug</Configurations> -->
+		<!-- <AnalysisLevelGlobalization>latest-recommended</AnalysisLevelGlobalization> -->
+		<!-- <TreatWarningsAsErrors>true</TreatWarningsAsErrors> -->
+		<!-- <NoWarn>RS0026</NoWarn> -->  <!-- Too many bogus warnings, see https://github.com/dotnet/roslyn-analyzers/issues/6757 -->
+	</PropertyGroup>
+
+	<PropertyGroup Condition=" '$(Configuration)' == 'Debug' Or '$(Configuration)' == 'FullDebug' ">
+		<DebugType>portable</DebugType>
+		<DebugSymbols>True</DebugSymbols>
+		<Optimize>false</Optimize>
+		<DefineConstants>$(DefineConstants);DEBUG;TRACE</DefineConstants>
+		<LibraryPKHash></LibraryPKHash>
+	</PropertyGroup>
+
+	<PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
+		<SignAssembly>True</SignAssembly>
+		<DelaySign>true</DelaySign>
+		<AssemblyOriginatorKeyFile>..\..\cql-sdk.publickey</AssemblyOriginatorKeyFile>
+		<IncludeSymbols>True</IncludeSymbols>
+		<GeneratePackageOnBuild>False</GeneratePackageOnBuild>
+		<SymbolPackageFormat>snupkg</SymbolPackageFormat>
+		<LibraryPKHash>00240000048000009400000006020000002400005253413100040000010001006d866cf0363db3976082a2439c420bd942b78243c0ce6c289f3c07d6901ecf149c38bafc12167f6eba7618b288e5327a7e4ab554c9a294ae6c68d7455b6562f846455b6bfdde68542a7046a632f1730a1647f2d365f8a25f73860c14d1428e7319df673e2d3212fa535f310c52a7c7ec17de772cb49961f4ea195fa4ea601dcd</LibraryPKHash>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<!--
+		JetBrains.Annotations help reduce false positive warnings, explicitly declare purity and nullability in your code,
+		deal with implicit usages of members, support special semantics of APIs in ASP.NET and XAML frameworks and
+		otherwise increase accuracy of JetBrains Rider and ReSharper code inspections.
+
+		All usages of JetBrains.Annotations attributes are erased from metadata by default, which means no actual binary
+		reference to 'JetBrains.Annotations.dll' assembly is produced. If you need to preserve these attributes in
+		metadata, just define 'JETBRAINS_ANNOTATIONS' conditional compilation symbol in your projects.
+		see more: https://www.nuget.org/packages/JetBrains.Annotations/
+		-->
+		<PackageReference Include="JetBrains.Annotations" Version="2024.2.0" />
+		<!-- <PackageReference Include="Microsoft.CodeAnalysis.PublicApiAnalyzers" Version="3.11.0-beta1.24318.1" PrivateAssets="All" /> -->
+	</ItemGroup>
+
+	<PropertyGroup>
+		<Box> __________________________________________________________________________________________&#10;/&#10;{Text}&#10;\</Box>
+	</PropertyGroup>
+
+</Project>
